Update processing to avoid floating point in counting categories.
A peculiar quirk of the Python pandas library we are using to
manipulate the tabular data, is that it uses floating-point numbers
to represent numbers in any column where nulls are present in some
rows.

With this commit I have added a bit of post-processing which converts
columns which should be treated as integers, to appropriately-formatted
strings, so the processed CSV output only has integers in those columns.